Updated CONTRIBUTING.md and README.md
In the CONTRIBUTING.md added FAQs, Feature request section, technical setup of DogeHouse on local computer for development. I moved it from the main README.md file because i thought it great to be placed in here since it's the contributing file for new contributions, and did a couple of gramma and english correction. In the README.md, cleared and added a couple of helpful links, so one can easily navigate to Contributing and code of conduct. Moved 'How to run locally' to the CONTRIBUTING.md file and added a direct link to that section. @benawad i think i'd be great to have that in there, as it's a file for contributors🙂